+++ app/src/main/java/woowacourse/payments/ui/newcard/util/transformation/CardNumberVisualTransformation.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.transformation
+package woowacourse.payments.ui.newcard.util.transformation
 
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.input.OffsetMapping
